rename Java package in jackson module from "gson" to "jackson"
