+++ grumpyjson-jackson/src/main/java/io/github/grumpystuff/grumpyjson/jackson/JacksonBasedJsonEngine.java
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-package io.github.grumpystuff.grumpyjson.gson;
+package io.github.grumpystuff.grumpyjson.jackson;
 
 
 import com.fasterxml.jackson.core.JsonParseException;
